Moved some files into their own separate directories.
