+++ src/openssf-dashboard/shared/models/scorecard-check-details.model.ts
@@ -1,0 +1,6 @@
+import { ScorecardCheck } from './scorecard-check.model';
+
+export interface ScorecardCheckDetails {
+  details: string
+  check: ScorecardCheck
+}
